requirements.txt
flake8==3.9.2
isort==5.12.0
python-dotenv==1.0.0
requests==2.28.2
pandas==2.0.0
hopsworks==3.0.5
joblib==1.2.0
scikit-learn==1.2.2
numpy==1.24.2
